requirements.txt
abc-py==0.0.1
absl-py==1.0.0
alabaster==0.7.10
antlr4-python3-runtime==4.8
asn1crypto==0.22.0
astroid==1.4.9
atomicwrites==1.4.1
attrs==22.1.0
Babel==2.4.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
chardet==3.0.3
charset-normalizer==2.0.4
click==6.7
cloudpickle==0.2.2
clyent==1.2.1
colorama==0.3.9
contextlib2==0.5.5
cryptography==38.0.4
cycler==0.10.0
Cython==0.29.27
dask==0.14.3
dataclasses==0.6
datashape==0.5.2
debugpy==1.5.1
decorator==4.0.11
distributed==1.16.3
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Cors==3.0.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.45.0
gtimer==1.0.0b5
HeapDict==1.0.0
html5lib==0.999
hydra-core==1.1.0
idna==2.5
imagesize==0.7.1
importlib-metadata==4.8.3
importlib-resources==5.10.2
ipdb==0.13.9
ipykernel==4.6.1
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isodate==0.6.1
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
joblib==1.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==6.4.2
jupyter-core==4.3.0
kiwisolver==1.4.4
lazy-object-proxy==1.2.2
locket==0.2.0
Markdown==3.3.6
MarkupSafe==0.23
matplotlib==3.3.4
mccabe==0.7.0
mistune==0.7.4
more-itertools==8.14.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numexpr==2.6.2
numpy==1.21.6
numpydoc==0.6.0
oauthlib==3.2.0
odo==0.5.0
olefile==0.44
omegaconf==2.1.2
openpyxl==2.4.7
packaging==16.8
pandas==1.1.0
pandocfilters==1.4.1
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.4.0
pluggy==1.0.0
ply==3.10
prompt-toolkit==3.0.36
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.5.1
py==1.11.0
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
pyg-lib==0.1.0+pt113cu116
Pygments==2.2.0
pylint==1.6.4
pyodbc==4.0.35
pyOpenSSL==17.0.0
pyparsing==2.1.4
PyQt5==5.9.2
PySocks==1.7.1
pytest==3.10.0
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==0.5.2
PyYAML==6.0
pyzmq==24.0.1
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
rdflib==6.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rope-py3k==0.9.4.post1
rsa==4.8
scikit-image==0.14.5
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.8
six==1.10.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.10.2
sympy==1.0
tables==3.7.0
tblib==1.3.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.6
testpath==0.3
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.13.1+cu116
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cu116
torch-sparse==0.6.16+pt113cu116
torchaudio==0.13.1+cu116
torchtext==0.14.1
torchvision==0.14.1+cu116
tornado==4.5.1
tqdm==4.63.0
traitlets==4.3.2
typing_extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.26.8
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
yacs==0.1.8
zict==0.1.2
zipp==3.6.0
